feat(cli): build --profile
