+++ packages/vite/src/node/cli.ts
@@ -1,3 +1,5 @@
+import path from 'node:path'
+import fs from 'node:fs'
 import { performance } from 'node:perf_hooks'
 import { cac } from 'cac'
 import colors from 'picocolors'
@@ -26,6 +28,27 @@ interface GlobalCLIOptions {
   m?: string
   mode?: string
   force?: boolean
+}
+
+export const stopProfiler = (log: (message: string) => void): void => {
+  // @ts-ignore
+  const profileSession = global.__vite_profile_session
+  if (profileSession) {
+    profileSession.post('Profiler.stop', (err: any, { profile }: any) => {
+      // Write profile to disk, upload, etc.
+      if (!err) {
+        const outPath = path.resolve('./vite-profile.cpuprofile')
+        fs.writeFileSync(outPath, JSON.stringify(profile))
+        log(
+          colors.yellow(
+            `CPU profile written to ${colors.white(colors.dim(outPath))}`
+          )
+        )
+      } else {
+        throw err
+      }
+    })
+  }
 }
 
 const filterDuplicateOptions = <T extends object>(options: T) => {
@@ -125,6 +148,7 @@ cli
       )
 
       server.printUrls()
+      stopProfiler((message) => server.config.logger.info(`  ${message}`))
     } catch (e) {
       createLogger(options.logLevel).error(
         colors.red(`error when starting dev server:\n${e.stack}`),
@@ -194,6 +218,7 @@ cli
       )
       process.exit(1)
     }
+    stopProfiler((message) => createLogger(options.logLevel).info(message))
   })
 
 // optimize
